Load lookups (when available) even if check_format=False
Change the behaviour of ExperimentListDict._load_pickle_path so that the
lookup files for mask, gain, pedestal and offset maps are loaded whenever
the file is available. Previously these were only loaded when
check_format=True. However, these files are produced by data processing
so normally they would be available alongside experiments and reflections
even if the raw images are not available. The existence of these
files does not really have anything to do with the Format.

This change ensures that dx, dy maps are loaded even when
check_format=False, which fixes https://github.com/dials/dials/issues/2744