FormatNXMX: Avoid storing duplicate static masks (#789)
When trying to dials.import a large number of experiment files, the RAM
usage on Linux was unnecessarily high due to separate static masks being
created for each file. Since masks tend to be limited in number for each
experiment, it makes more sense to store masks with unique values only,
which is what this PR aims to do.

Closes https://github.com/dials/dials/issues/2227.