Tidy up consolidation of histories
This is better as a method of the ExperimentList class, providing
an easier interface to history management. When you want to append a
new history line to each experiment, just call el.consolidate_histories(),
which gives you the now unique History object attached to each experiment
and append the line to that object.