In the refactor I inadvertently lost the changed definition of the goniometer axes for the ED1. Put that back in here, but note I suspect there is still something incorrect about the geometry model. In particular, for Photon-II/III the slow axis is defined as 0,1,0, but this is not like the older FormatBruker detectors, or like how it is set for the ED devices. I think slow along +Y also diasgrees with the Bruker Frame Format documentation, which states:
The order of the pixels in the file corresponds to the display of pixels on a screen, like the raster-scan
order of a CRT display. The first pixel in the file corresponds to the upper-left corner of the display, the
512th to the upper right, and the last pixel to the lower right. This data ordering convention is the one
used by most display devices and by the X-Windows library. Frames display as if viewed from the source
toward the detector.