Tidy dxtbx.serialize.filename.resolve_path
This is used in several places. Annotate, but also make explicitly handle
pathlib.Path objects, including returning a Path if the function is passed a
Path in the first place.

Technically there is a minor change in the outer API: Previously, if the
function was passed None as the path=, it would change the return type to
"" (empty string). Now, passing None is forbidden by the type annotations - but
if a None is passed in anyway, it will return None (the same type).