Longtime bugfix for averaging
Indentation error here meant a factor of n panels was added to the denominator of averaging data.  But because that same factor was added to the numerator for distance and wavelength, the bug wasn't evident in the geometry of the average.  But for multipanel detectors the average and standard deviation were affected by this factor of n panels.