LCLS psana2 support (#860)
One of three pull requests to add psana2 support to cctbx.xfel and dials. The others are:

cctbx/cctbx_project#1111
dials/dials#3038

This is a substantial change to FormatXTC and its derivatives and it includes edits to FormatMultiImage and even experiment_list.py. Most of the changes are to account for the following:

- psana2 requires the use of MPI, and it specifically reserves ranks 0 and 1. Every rank must import psana for initialization, and then every rank must load the data. Ranks 0 and 1 will have no data, however they are used to coordinate the data read, so any MPI program must use them to read the data, then not use them again.
- Because ranks 0 and 1 will have no data, changes to FormatMultiImage::get_imageset() and ExperimentListFactory:: from_stills_and_crystal are need to allow zero-length imagesets
- The new psana2 API is adopted
- The code is cross-compatible between psana1 and psana2

Includes a fix for dxtbx.image_average computing the average and standard deviation images

---------

Co-authored-by: Frédéric P. Poitevin <frederic.poitevin@gmail.com>
Co-authored-by: Monarin Uervirojnangkoorn <monarin@gmail.com>
Co-authored-by: Daniel W. Paley <dwpaley@lbl.gov>
Co-authored-by: Derek Mendez <dermen@lbl.gov>
Co-authored-by: Christopher O'Grady <cpo@slac.stanford.edu>